+++ packages/core/modules/utils/autocomplete.js
@@ -1,10 +1,13 @@
 import {sleep} from "./stuff";
-import {listValuesToArray, mapListValues} from "./listValues";
+import {listValuesToArray, getListValue, makeCustomListValue} from "./listValues";
 
 export const simulateAsyncFetch = (all, cPageSize = 0, delay = 1000) => async (search, offset, meta) => {
   const pageSize = meta?.pageSize != undefined ? meta.pageSize : cPageSize;
   const filtered = listValuesToArray(all)
-    .filter(({title}) => search == null ? true : title.toUpperCase().indexOf(search.toUpperCase()) != -1);
+    .filter(({title, value}) => search == null ? true : (
+      title.toUpperCase().indexOf(search.toUpperCase()) != -1
+      || `${value}`.toUpperCase().indexOf(search.toUpperCase()) != -1
+    ));
   const pages = pageSize ? Math.ceil(filtered.length / pageSize) : 0;
   const currentOffset = offset || 0;
   const currentPage = pageSize ? Math.ceil(currentOffset / pageSize) : null;
@@ -23,31 +26,59 @@ export const simulateAsyncFetch = (all, cPageSize = 0, delay = 1000) => async (s
   };
 };
 
-export const mergeListValues = (values, newValues, toStart = false) => {
+export const mergeListValues = (values, newValues, toStart = false, hideNewValues = false) => {
   if (!newValues)
     return values;
   const old = values || [];
-  const newFiltered = newValues.filter(v => old.find(av => av.value == v.value) == undefined);
+  const newFiltered = newValues
+    .filter(v => old.find(av => ""+av.value == ""+v.value) == undefined)
+    .map(v => (hideNewValues ? {...v, isHidden: true} : v));
   const merged = toStart ? [...newFiltered, ...old] : [...old, ...newFiltered];
   return merged;
 };
 
+export const optionToListValue = (val, listValues, allowCustomValues) => {
+  const v = val == null || val == "" ? undefined : (val?.value ?? val);
+  const item = getListValue(v, listValues);
+  const customItem = allowCustomValues && !item ? makeCustomListValue(v) : undefined;
+  const listValue = item || customItem;
+  const lvs = listValue ? [listValue] : undefined; //not allow []
+  return [v, lvs];
+};
+
+export const optionsToListValues = (vals, listValues, allowCustomValues) => {
+  const newSelectedListValues = vals.map((val, _i) => {
+    const v = val == null || val == "" ? undefined : (val?.value ?? val);
+    const item = getListValue(v, listValues);
+    const customItem = allowCustomValues && !item ? makeCustomListValue(v) : undefined;
+    const listValue = item || customItem;
+    return listValue;
+  }).filter(o => o != undefined);
+  let newSelectedValues = newSelectedListValues
+    .map(o => (o?.value ?? o));
+  if (!newSelectedValues.length)
+    newSelectedValues = undefined; //not allow []
+  return [newSelectedValues, newSelectedListValues];
+};
+
 export const listValueToOption = (lv) => {
   if (lv == null) return null;
-  const {title, children, value, disabled, groupTitle, renderTitle} = lv;
-  let option = {title, value};
-  if (!title && children)
-    title = children
+  const {title, value, disabled, groupTitle, renderTitle, children, label, isCustom, isHidden} = lv;
+  let option = {
+    value,
+    title: title || label || children, // fix issue #930 for AntD
+  };
   if (disabled)
     option.disabled = disabled;
   if (groupTitle)
     option.groupTitle = groupTitle;
   if (renderTitle)
     option.renderTitle = renderTitle;
+  if (isCustom)
+    option.isCustom = isCustom;
+  if (isHidden)
+    option.isHidden = isHidden;
   return option;
 };
 
-export const getListValue = (selectedValue, listValues) => 
-  mapListValues(listValues, (lv) => (lv.value === selectedValue ? lv : null))
-    .filter(v => v !== null)
-    .shift();
+export { getListValue };