+++ CHANGELOG.md
@@ -10,6 +10,9 @@
   - Fix totalQuery structure in mongodb (PR #916) (issue #915)
   - Fixed bug in `loadTree()` (PR #917) (issue #356)
   - Fixed bug on change group value (PR #944) (issue #923)
+  - Fixed render of option titles in Autocomplete AntDesign (PR #947) (issues #930, #942)
+  - Implemented `allowCustomValues` for MUI (PR #947) (issue #327)
+  - Improved autocomplete for MUI and AntD. Mark custom values with coral color (PR #947)
 - 6.2.0
   - Fixed type `Config`: should have render settings like `renderSize` (PR #909) (issue #879)
   - Fixed type for `renderBeforeWidget`: `RuleProps` instead of wrong `FieldProps` (PR #909) (issue #879)